JS Logger: Use new endpoint, ES6, use tracekit (#6672)
* Move logger to wepack build flow

* Remove js error logger chunk

* Remove supporting old windowonerror - should play nicely with hto reload

* Add tracekit func to all of that

* Add tracekit to package.json

* Update shrinkwrap
