+++ package.json
@@ -116,6 +116,7 @@
     "superagent": "1.2.0",
     "tinymce": "4.3.12",
     "to-title-case": "0.1.5",
+    "tracekit": "0.4.3",
     "tv4": "1.2.7",
     "tween.js": "16.3.1",
     "twemoji": "1.3.2",
